app/requirements.txt
azure-ai-projects 
azure-identity
python-dotenv
semantic-kernel[azure]
chainlit
azure-storage-blob